move compare fns to IAMAPI, verify_capabilities and compare_capabilities
