+++ tests/tests_unit/test_data_classes/test_capabilities.py
@@ -308,9 +308,9 @@ class TestProjectCapabilityList:
         ],
     )
     def test_has_capability(
-        self, proj_capabs_list: ProjectCapabilityList, project_name: str, capability: Capability
+        self, cognite_client, proj_capabs_list: ProjectCapabilityList, project_name: str, capability: Capability
     ) -> None:
-        assert not proj_capabs_list.compare([capability], project=project_name)
+        assert not cognite_client.iam.compare_capabilities(proj_capabs_list, [capability], project=project_name)
 
     @pytest.mark.parametrize(
         "capability",
@@ -323,20 +323,26 @@ class TestProjectCapabilityList:
         ],
     )
     def test_is_missing_capability(
-        self, proj_capabs_list: ProjectCapabilityList, project_name: str, capability: Capability
+        self, cognite_client, proj_capabs_list: ProjectCapabilityList, project_name: str, capability: Capability
     ) -> None:
-        missing_acls = proj_capabs_list.compare([capability], project=project_name)
+        missing_acls = cognite_client.iam.compare_capabilities(proj_capabs_list, [capability], project=project_name)
         assert missing_acls == [capability]
 
-        missing_acls = proj_capabs_list.compare([capability], project="do-es-nt-ex-is-ts")
+        missing_acls = cognite_client.iam.compare_capabilities(
+            proj_capabs_list, [capability], project="do-es-nt-ex-is-ts"
+        )
         assert missing_acls == [capability]
 
-    def test_partly_missing_capabilities(self, proj_capabs_list: ProjectCapabilityList, project_name: str) -> None:
+    def test_partly_missing_capabilities(
+        self, cognite_client, proj_capabs_list: ProjectCapabilityList, project_name: str
+    ) -> None:
         has = RawAcl([RawAcl.Action.Read], scope=RawAcl.Scope.Table({"my_db": ["my_table"]}))
         has_also = EventsAcl([EventsAcl.Action.Write], scope=EventsAcl.Scope.DataSet([1, "2"]))
         has_not = RawAcl([RawAcl.Action.Read], scope=RawAcl.Scope.Table({"my_db": ["unknown_table"]}))
 
-        missing_acls = proj_capabs_list.compare([has, has_not, has_also], project=project_name)
+        missing_acls = cognite_client.iam.compare_capabilities(
+            proj_capabs_list, [has, has_not, has_also], project=project_name
+        )
         assert missing_acls == [has_not]
